r10462/sofa : Fix "loss of data" warnings in extlibs
Former-commit-id: e32d21e7dcbb70f4595070183fa56ffbe872bcb9

+++ extlibs/miniBoost/boost/graph/bandwidth.hpp
@@ -51,8 +51,8 @@ namespace boost {
     vertices_size_type b = 0;
     typename graph_traits<Graph>::edge_iterator i, end;
     for (boost::tie(i, end) = edges(g); i != end; ++i) {
-      int f_i = get(index, source(*i, g));
-      int f_j = get(index, target(*i, g));
+      int f_i = (int)get(index, source(*i, g));
+      int f_j = (int)get(index, target(*i, g));
       b = max BOOST_PREVENT_MACRO_SUBSTITUTION (b, vertices_size_type(abs(f_i - f_j)));
     }
     return b;
