+++ extlibs/newmat/newmat/myexcept.cpp
@@ -74,7 +74,7 @@ void BaseException::AddMessage(const char* a_what)
 {
    if (a_what)
    {
-      int l = strlen(a_what); int r = LastOne - SoFar;
+      int l = (int)strlen(a_what); int r = LastOne - SoFar;
       if (l < r) { strcpy(what_error+SoFar, a_what); SoFar += l; }
       else if (r > 0)
       {
